update dashboard and questions page
I have updated the dashboard page. The name of the topic and the name of
the lecturer can seen. When the topic is clicked, you are redirected to
the page where the questions under that topic are on another page.

I have also updated the Questions page to allow a user select an answer,
show the number of question they are currently on and go to the next
question when they click on the next button under the same topic.